Module aws_sdk_ec2instanceconnect::error

Modules

auth_exception

See AuthException

ec2_instance_not_found_exception

See Ec2InstanceNotFoundException

ec2_instance_type_invalid_exception

See Ec2InstanceTypeInvalidException

invalid_args_exception

See InvalidArgsException

serial_console_access_disabled_exception

See SerialConsoleAccessDisabledException

serial_console_session_limit_exceeded_exception

See SerialConsoleSessionLimitExceededException

serial_console_session_unavailable_exception

See SerialConsoleSessionUnavailableException

service_exception

See ServiceException

throttling_exception

See ThrottlingException

Structs

AuthException

Either your AWS credentials are not valid or you do not have access to the EC2 instance.

Ec2InstanceNotFoundException

The specified instance was not found.

Ec2InstanceTypeInvalidException

The instance type is not supported for connecting via the serial console. Only Nitro

InvalidArgsException

One of the parameters is not valid.

SendSSHPublicKeyError

SendSerialConsoleSSHPublicKeyError

SerialConsoleAccessDisabledException

Your account is not authorized to use the EC2 Serial Console. To authorize your

SerialConsoleSessionLimitExceededException

The instance currently has 1 active serial console session. Only 1 session is supported at a time.

SerialConsoleSessionUnavailableException

Unable to start a serial console session. Please try again.

ServiceException

The service encountered an error. Follow the instructions in the error message and try again.

ThrottlingException

The requests were made too frequently and have been throttled. Wait a while and try again.

Enums

SendSSHPublicKeyErrorKind

SendSerialConsoleSSHPublicKeyErrorKind